Server/requirements.txt
Flask==3.0.3
requests==2.32.3

cryptography~=43.0.0
python-dotenv~=1.0.1
waitress~=3.0.0
bcrypt~=4.2.0